% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/218ChisqLRtest.R
\name{chisq.LRtest}
\alias{chisq.LRtest}
\title{Chi-Squared Liklihood Ratio Test}
\usage{
chisq.LRtest(tab)
}
\arguments{
\item{tab}{An I x J contingency table passed through \code{table()}.}
}
\value{
Returns the likelihood ratio test statistic, G^2, the p-value of the test, and the degrees of freedom,
    (I-1)(J-1).
}
\description{
This function calculates the likeliehood ratios for the independence test of the same name.
}
\examples{
# make some table (for the purposes of this example, 2x2)
table <- matrix(c(5134,2829,3946,715), nrow=2, ncol=2, byrow=TRUE)
# then pass this table through chisq.LRtest
chisq.LRtest(table)
}
