% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalOddsRatioPackage.R
\name{local.odds}
\alias{local.odds}
\title{Local Odds Ratio}
\usage{
local.odds(tab)
}
\arguments{
\item{tab}{An I x J contingency table passed through \code{table()}.}
}
\value{
\code{local.odds()} returns a matrix of the local odds ratios for the provided table
}
\description{
Calculates the local odds ratios for an inputted I x J table
}
