% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/218OddsRatio.R
\name{odds}
\alias{odds}
\title{Odds Ratio}
\usage{
odds(tab, conf.level = 0.95)
}
\arguments{
\item{tab}{A 2 x 2 contingency table passed through \code{table()}.}

\item{conf.level}{The confidence level you are testing at. Defaults to 95\%}
}
\value{
Returns the odds ratio test statistic, theta, the confidence level
    of the interval, and the confidence interval itself.
}
\description{
Calculates the odds ratios for an inputted 2 x 2 table
}
\examples{
# create some 2x2 contingency table 
table <- matrix(c(766,110,702,89), nrow=2, ncol=2, byrow=TRUE)
# pass it through odds at default confidence level of 0.95
odds(table)
# change confidence level to reflect desired
odds(table, conf.level=.90)
}
