% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearTrendPackage.R
\name{linear.trend}
\alias{linear.trend}
\title{Linear Trend Test}
\usage{
linear.trend(x, row.scores = NULL, col.scores = NULL)
}
\arguments{
\item{x}{A 2 x 2 contingency table of structure matrix,
or factor (the latter being passed through as \code{table()})}

\item{row.scores}{Numerical value assigned to rows for calculation purposes; Defaults to Null}

\item{col.scores}{Numerical value assigned to columns for calculation purposes; Defaults to Null}
}
\value{
The M^2 test statistic, the parameters for the test (r and df), the type of
    test performed, and the resultant p-value.
}
\description{
This function performs a Mantel Haenszel Chi-Square Test for Two Way Tables.
}
